This is NOT FIT FOR MERGE.
This is a first pass hack to store the processed sensor list in local storage.

It does not have sufficient error handling
It does not expire the sensor list ever
The app will likely totally break for end users if bad data is stored in sensors.

But maybe it's a useful starting point for someone who really knows javascript.